backend/ruralsiksha/urls.py
"""
URL configuration for ruralsiksha project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.2/topics/http/urls/
"""
from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView

urlpatterns = [
    path('admin/', admin.site.urls),

    # API docs
    path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
    path('api/docs/', SpectacularSwaggerView.as_view(url_name='schema'), name='docs'),

    # JWT Authentication
    path('api/auth/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/auth/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    # App URLs
    path('api/', include([
        path('users/', include('users.urls')),
        path('resources/', include('resources.urls')),
        path('quizzes/', include('quizzes.urls')),
        path('ai-tutor/', include('ai_tutor.urls')),
        path('doubts/', include('doubts.urls')),
        path('notifications/', include('notifications.urls')),
    ])),
]
