backend/requirements.txt
Django>=4.2,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
requests>=2.31
python-dotenv>=1.0
drf-spectacular>=0.26
